{{cookiecutter.project_name}}/templates/sagemaker/requirements.txt
aina-gradio-theme==1.0
gradio=={{cookiecutter.gradio_version}}
python-dotenv==1.0.0
